Move scheme initialization to Vuex
